+++ packages/dht/src/rpc-protocol/DhtRpcOptions.ts
@@ -7,5 +7,5 @@ export interface DhtRpcOptions extends ProtoRpcOptions {
     clientId?: number
     connect?: boolean
     sendIfStopped?: boolean
-    doNotBufferWhileConnecting?: boolean
+    bufferWhileConnecting?: boolean
 }