+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changes before Tatum release are not documented in this file.
 - Add `StreamrClient#publishRaw()` for publishing raw messages (https://github.com/streamr-dev/network/pull/3280)
 - Add new `keys` configuration to the `encryption` section (https://github.com/streamr-dev/network/pull/3284)
 - Add new `validation` configuration section (https://github.com/streamr-dev/network/pull/3302)
+- Add new configuration options for controlling content delivery buffering (https://github.com/streamr-dev/network/pull/3305)
 
 #### Changed